Simplify processedText
