+++ packages/shared/src/utils/regex.ts
@@ -22,3 +22,6 @@ export const HTTPS_URL = /^https?:\/\//i;
 
 // check if ethereum address
 export const ETH_ADDRESS = /^0x[a-fA-F0-9]{40}$/;
+
+// check markdown link
+export const MARKDOWN_LINK_REGEX = /\[([^\]]+)]\((https?:\/\/[^\s]+)\)/g;
